2010-08-31  Paolo Carlini  <paolo.carlini@oracle.com>
	* testsuite/23_containers/forward_list/operations/remove_freed.cc:
	Fix test01 return type to void.
	* testsuite/util/exception/safety.h: Avoid -Wall -m32 warnings.
	* testsuite/util/replacement_memory_operators.h: Likewise.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@163684 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libstdc++-v3/testsuite/util/exception/safety.h
@@ -52,13 +52,13 @@ namespace __gnu_test
 	  __s += "\n";
 	  __s += "random number generated is: ";
 	  char buf[40];
-	  __builtin_sprintf(buf, "%lu", random);
+	  __builtin_sprintf(buf, "%lu", (unsigned long)random);
 	  __s += buf;
 	  __s += " on range [";
-	  __builtin_sprintf(buf, "%lu", distribution.min());
+	  __builtin_sprintf(buf, "%lu", (unsigned long)distribution.min());
 	  __s += buf;
 	  __s += ", ";
-	  __builtin_sprintf(buf, "%lu", distribution.max());
+	  __builtin_sprintf(buf, "%lu", (unsigned long)distribution.max());
 	  __s += buf;
 	  __s += "]\n";
 	  std::__throw_out_of_range(__s.c_str());
